Initial commit
First draft of pseudonymisation script.

Replaces any potentially user-identifying info that we need to
keep with a pseudonym. Drops information that we do not need.
Expands out information nested inside JSON blobs into new columns.
Tried with meetings workflow of Scarlets and Blues and with all
workflows of HMS NHS.