Better JSON dumping
The trouble with the previous approach is that, if a later key
contains a non-dict value where an earlier key contained a
dict value, we lose the knowledge of those keys.

This approach is completely additive, so it will catch every key
that is in a dict. It still won't show us keys in dicts embedded
in other data types (e.g. a list of dicts).

This seems good enough to give us a sense of the kind of data that
is in the JSON.