Handle same JSON field in different 'spellings'
So far the only alternative spellings are 'priority' for
'#priority' and various bits of RBGE subject metadata that begin
with a capital or a lowercase.

This code handles the case where we have two spellings of the same
field and one of the spellings is the all-lowercase version of the
other one. Could generalise it further if this turns out to be
needed.

'#priority' vs 'priority' is ignored but allowed through the
validation check. Could merge these columns if that turns out to
be needed.