HACK: Aggregate the line charts
The benefit of this approach is that it is minimally-
invasive, minimising the risk of messing up anything
that I already have in place.

Stop multiprocessing the by-workflow charts so that I can
get return values back from them without figuring out
shared variables right now.

Draw by project for all projects

Also draw by (a special version of) workflow type for HMS NHS.

This is just a hack to generate charts for the reports. Better
solution: dumping the underlying data of the charts to
CSV files (which I should be doing anyway), then reading them
back in to generate my chart. Another better solution: learning
to set up a shared variable and keeping the multiprocessing in
place.