Tighten datetime usage
Keep datetimes in timezone-naive format, As all datetimes are in
UTC anyway, this should make no difference at all, but it does
remove one complexity from thinking about this.

Also change the unit specifier for putting utc_offset into the
appropriate type from 's' to 'seconds' -- 's' appears to work, but
the documentation does not have it as a legal value here.