Write identities file in a readable, diffable fashion
Setting indent makes the file human readable, and adds line breaks
which are helpful for diff tools.

Sorting the keys makes sure that all dictionaries come out in the
same order, which can also be helpful when diffing.

Unrelatedly, rather than write a json.dumps output to a file, just
use json.dump directly. It works in the same way, but outputs to
file intead or returning a string.