Generate random pseuondyms
Previously we were incrementing a counter. This seems a touch
better. As a happy side effect, we get pseudonyms of constant
length.

Could use the alphabet here to get still shorter pseudonyms, but
perhaps it is neat to stick to digits. They are less likely to
accidentally resemble real usernames, too -- though I think it
is already pretty unlikely that we'll generate anything
resembling a real username.