Only split out the branches of meetings for all_classifications.csv
I'm not a huge fan of this change, as it introduces a requirement
that we not drop any rows before we do the split-out. However,
this is currently the case, so it is OK.

A smarter way to do this would be to use the subjects file to
identify the attendance pages, as we have not modified that.