Add replication recipes to pseudonymise
Generate READMEs with some instructions for each CSV file.

For the per-project files, which exist only for the data sharing
platform, add these READMEs and the CSV file to an archive file
in a new sharing directory.

Generate both Windows and *nix friendly archive formats.

Add a new script, distribute.py, to extend the readme for
all_classifications to include graph regeneration. Add this readme,
all_classifications.csv, all HTML graphs and all CSV inputs
giving the source data for the graphs, as well as the run log, to
another zip/tarball in the sharing directory.