Improve location generation
Drop the read_subj_locs function -- most of the time we can piece
the subject location together using information in the
classifications file. This might be less efficient, but it makes
for simpler and more accurate code.

Add more locations -- locations that drop into the indexer, for
better indexed access, and institution-specific location for
TNA.

Also fix bugs in the code to read in the subjects file. We now
only need this file to get the subject_set_id, for the one case
where we need that to construct a URL.

And warnings for cases where a location of a given type cannot
be found.

Larger change than I like to commit in one go, apologies.