Fix blurb forks for different kinds of README
Have mistakenly been assuming that a str parameter indicates a
README for a single project, while non-str parameter indicates a
README for all projects.

Actually the parameter is a list and what matters is its length.

Factor out a variable to tell me what kind of README I am writing
and set it based on length of list parameter, assert-faulting if
the length is surprising.

While I'm there, reorder the if statements to look tidier.