Monochrome heatmaps
Make the heatmaps greyscale. By being monochrome we improve
accessibility. Greyscale seems the simplest way to do this.

Use some trickery to make the heatmap-histograms patterned
rather than coloured. This makes them stand out clearly from the
background, without occupying the same "semantic space" as the
colour density in the heatmap.

While this trickery might be a layering violation (see comment),
it appears to work well enough to generate improved charts for the
final report. It is very encouraging that I get an error if I
try to insert a key or value into these dictionaries that is not a
part of the defined interface.