Update README to explain about variations in subjects.
In a bit more detail:
The data that we generate is based on the Zooniverse exports.
We do not version control these files.
For the 'classifications' exports this should be fine -- we have
cut-off datestamps and workflow versions to use to control our
input. So far as I know, new classifications will not get added to
our input, so we can always re-download exports and get the same
results.
The subjects file is a different story. If the subjects are
changed (e.g. removed) then the URLs will of course change.