Transfer most remaining project-specific stuff to data.py
This will allow using the scripts in a new configuration by just
swapping in a different version of data.py.

The single all-classifications file is left in place, just
becuase I don't need to abstract this away at the moment -- a
later patch will add a CLI switch to just not generate this file.